deleting shapeBreeding files and moving funcs to shapeUtils
