+++ kushekbaev.danil/T1/shapeUtils.hpp
@@ -6,6 +6,11 @@
 
 namespace kushekbaev
 {
+  namespace details
+  {
+    bool isTriangle(const point_t first, const point_t second, const point_t third);
+    bool isPointInsideTriangle(const point_t first, const point_t second, const point_t third, const point_t last);
+  }
   void movePoints(point_t** points, const size_t size, const double dx, const double dy);
   void scalePoints(point_t** points, const size_t size, const double scaleCoeff, const point_t mid);
   double getLineLength(const point_t first, const point_t second);
